Fix: double->float 변경

+++ backend/incense/src/main/java/com/suyang/incense/api/response/analysis/CloudDto.java
@@ -16,7 +16,7 @@ public class CloudDto {
     @Setter
     public static class WordDto{
         private String text;
-        private double value;
+        private float value;
 
     }
